Consolidate sort proxy into views/partials/sort_proxy.py; drop SizeItem
Extract one shared SortProxyModel used by both Archive and Source tables,
removing the duplicate ArchiveSortProxyModel; fix both-None strict-weak
ordering. Delete dead SizeItem + orphaned imports, drop stale
SIZE_DECIMAL_DIGITS import, make updateThreads an instance attribute.
